PI-1013 updated to use merge endpoint for reschedule also
